Move structs to their own file
